Added tests for creating and validting new customer - tests passing

+++ src/main/java/com/company/customermanagementservice/service/CustomerServiceImpl.java
@@ -4,7 +4,9 @@ import com.company.customermanagementservice.model.Customer;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityNotFoundException;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class CustomerServiceImpl implements CustomerService {
@@ -17,6 +19,7 @@ public class CustomerServiceImpl implements CustomerService {
 
     @Override
     public Customer createCustomer(Customer customerToBeCreated) {
+
         return customerRepository.createCustomer(customerToBeCreated);
     }
 
@@ -28,6 +31,12 @@ public class CustomerServiceImpl implements CustomerService {
     @Override
     public void deleteCustomer(Long customerId) {
 
-        customerRepository.delete(customerId);
+        Optional<Customer> customer = customerRepository.findById(customerId);
+
+        if(customer.isPresent()){
+            customerRepository.delete(customerId);
+        }else{
+            throw new EntityNotFoundException(String.format("Customer with Id %s does not exist", customerId));
+        }
     }
 }
